requirements.txt
Flask~=3.0.3
pillow~=10.4.0
google-generativeai~=0.7.0
google-api-python-client~=2.139.0
wikipedia~=1.4.0
requests~=2.32.3